Make setup_stack aware of Heat concurrency issue
This is a workaround for a concurrency issue detected on Heat
OSP-14, but that could be affecting any version of Heat (no
investigation has been perfomed). It is unknow if this is a bug
or just a limitation of the service.

The issue:

Heat looks like has a concurrency issue when creating new stacks with
the same name on ame time from more parallel processes.

When Tobiko Neutron test cases are executed using tox -e neutron command
test case execution is spawn along some worker processes.

It happens that the same HeatStackFixture, as it is shared between more
test cases, is being created on the same time from more workers.

The stack creation request is therefore submitted to Heat service with
the same parameters and despite what understood from the documentation,
more than one stack starts being created on the same time with the same
name.

Workaround implementation

After these stack creation is started tobiko asks for stack status
using stack name and all worker processes receives information of
the same stack (with the same stack ID).

Therefore those workers that creates a new stack with an ID different
from the one got by the stack name are deleted, leaving on most of the
cases the cloud with only one stack instance for the same stack name.

Change-Id: Ibcdc04d9436664b788d2ed6d68e2a20c74bd6147