Modify quick-start doc
Quick-start, not surprisingly, should be a quick start guide for the
user who tries Tobiko. Right now I'm afraid it's the opposite. It
deep dives into details that should be somewhere else but not in
the quick start page.

Also, most of the information written there is no longer relevant.
I have changed it to include the minimal instructions the user needs
for a true quick start experience with Tobik.

Change-Id: I0630aeeb736e51cab5e66750c4c8a726c1195af5