Cleanup zuul configuration
Remove publishing jobs, this project is not under governance
and cannot publish to docs.openstack.org. Replace the publishing
jobs by jobs that just test but do not publish.

Remove name from project stanza, it is not needed and will only
break in case of renames.

Use openstack-cover-jobs template instead of individual job.

Change-Id: I187d4ae7150ee9b4888af70e54cf6f2d3f80225a