Fix build autogenerated config files doc on RTD
This commit changes working dir from which sphinx-build is
run when using tox to make it working in same way as it is
done on readthedocs.org when documentation is build.

This commit also fixes path to oslo-config-generator config file
that it will be fine for spinx-build to autogenerate config.

Change-Id: I14982b3cbaa0bc87390e8694c4c866a3b7c375fa