Create Ansible role to install arbitrary Python version
The purpose of this is to be able to install a version
of python that is not available on target node. It
downloads source code from Python site, compiles it
using default compiler and install it.

It also installs and upgrade python packages (pip,
setuptools, etc.) and install additional ones required
by user via Pip

Change-Id: I3212d6fc1872451942aaa82e613236351f8592a3