Skip tests from RebootNodesTest on non TripleO based cloud
Those tests are designed to be run on TripleO based cloud as
requires undercloud node and overcloud nodes.
So lets skip them in e.g. devstack based deployments to make
job in overall to be green.

It also renames this test class to be "RebootTripleoNodesTest"
to better express what are those tests doing.

Change-Id: I0a4e710a14648248c53ba357a6a461bf8da2853f