Add way to know topology is based on containers
It is important in some cases to know if cloud topology is
contenerized (like TripleO) or not (like e.g. Devstack).
In case of contenerized topology, we need for now to skip
radvd related faults test as this isn't run in side car container
so it's gone always when neutron-l3-agent is stopped.

Change-Id: If7f93c21814cfc75cdd3d32900403dc978b53afe