Fix looking for ovs agent based on compute node name
In neutron agents faults tests module, there is test which looks
for neutron-ovs-agent which runs on same host as VM is spawned to
stop this neutron-ovs-agent.
This usually worked fine, but in case e.g. of TripleO hostname
given in neutron agents list contained full domain name and
hostname from nova's output containes short name.
Because of this missmatch there was no neutron-ovs-agent found to
stop and test was failing.
This patch fixes this issue by changing comparison of hostnames
of vm's host and agent's host.

Additionally this patch adds proper handling of the case when
neutron-ovs-agent isn't found.

Change-Id: If240188392feb6067af4f9a4a82e8ed95b4b99ee