Fix build of tobiko docs
Patch [1] moved "pandas" lib from requirements to the extra-requirements
to make it optional dependency.
That broke builing of the Tobiko docs as pandas is still needed in some
Tobiko code and it isn't installed to build docs anymore.

So lets fix that by including extra-requirements in the
doc/requirements.txt file too.

Change-Id: Ia55c0022fc4ec69873ea1c5667eec9685b5f3fb3