Use default logging format in Tobiko logs
Devstack's function "setup_logging" is setting format of log lines
to be correct in systemd and that works for services like Nova or
Neutron well in the CI when Devstack is used.
But for Tobiko logs it logs entries without timestamps which sometimes
can make debugging of the issues harder.

So this commit is removing call to setup_logging function in the
Tobiko's devstack plugin. Lets use default log format provided by the
oslo_log.

Change-Id: I4930d103b331258852b94243f61791e532b1c048