Add faults test which destroys and creates physical bridges
Test is destroying physical bridges on the node where router is
hosted so that floating IP shouldn't work at all.
Then it creates bridges again and checks if neutron-ovs-agent
was able to properly recover from that and FIP will work again.

Unfortunatelly this test can't work in e.g. devstack based u/s jobs
as there is additional bridge br-infra which provides connection between
nodes and this br-infra bridge is connected with br-ex by patch port.
When tobiko deletes br-ex and creates it again, there is this patch port
to br-infra missing and test is failing alwasy.

Test should works fine on deployments based on TripleO.

Change-Id: I125285f5f8abd2b2a285a7e46c954d6180d3ec32