Confirm node is rebooted when reboot command is sent
Some tobiko cloud disruption tests are unstable because once the reboot
command is sent to an overcloud node, that node status is validated
immediately. Sometimes, the reboot has not been applied yet.

With this patch, the node uptime is validated after the reboot command
is sent in order to guarantee that the reboot has been completed

Change-Id: I95607924c204744674f7fa08a74dc895ddc58f89