Fix connectivity issue during external bridge recreation test
In the faults test where we are testing if neutron-openvswitch-agent
can properly recover after deletion and recreation of the external
bridges, we are deleting all bridges which are in bridge-mappings for
the neutron-openvswitch-agent.
As we are doing that usually for an agent which runs on controller, we
may lost connectivity to the OpenStack api as well. Better to save ip
address to ping before the test to not to query API when it might be
not accessible.

Change-Id: Ieb9a59223639e3bbd997c6af748593d21cf8cba7