Openstack CLI tests failed on devstack
The issue is happening because of the devstack is using secure
connection for all the API calls with the self signed certificate.
This certificate is not trusted by the host so all the openstack
client tests are failed. Need to specify the certificate file with
the `--os-cacert` command line parameter to trust the certificate.

Change-Id: I3f2606d1da5fbcf1bf6ec1472748ddeb6906dd79