Fix skip of faults OVN agents tests when there is no OVN in env
Faults tests related to the OVN agents should be skipped in case when
Tobiko is run on the env without OVN, like e.g. ML2/OVS.
It wasn't skipped because super() method of the OVN test cases class was
called after some method which was looking for container name.
Now it is called in correct order and tests are skipped properly.

Change-Id: I0616169f91cb8ec1431ac91d34c1b2b20466f85a