Add JournalLogDigger class to parse OpenStack services logs
When we use devstack based environment, services like Nova or
Neutron are logging to the journalctl instead of files.
So we need to have som LogDigger class which will be able to parse
such logs in same way like log files.

This patch adds such class and also adds to the topology class
parameters which defines what LogDigger class should be used for
the topology and what are service names to log file names mappings.

Change-Id: Ice25bfa0cae60dc2b5402e59f549e2287808ba87