Remove installation of tobiko in the doc/source/conf.py file
Docs configuration file doc/source/conf.py shouldn't install
dependencies, like tobiko project itself as it seems that
this causes failures of the docs builds on the readthedocs.

So this patch removes it from that file and adds sphinx_rtd_theme
to the docs requirements file.
It also moves installation of tobiko project to the tox.ini file to make
Zuul docs job happy.

Change-Id: I0b227f87ade2d9561dd8aee988a565cbef31c7ae