Ensure that VM needed by the OvnController faults tests really works
In case when VM wasn't created earlier by some other tests and only
OvnController tests is run, it could fail due to vm which is in ERROR
state. And it's in such state due to already dead ovn-controller agent.

To avoid that this patch adds check if vm is really ready and reachable
before starting killing processes.

Change-Id: I6455354ef8686a12b3d105660398bef6db653e8d