Generalize and refactor Octavia waiters module
So far tobiko/tobiko/tests/scenario/octavia/waiters.py could only
wait for provisioning_status and for operating_status, and only for
"ACTIVE" status.

This patch generalizes the waiters module so it could wait for any
Octavia object, for any status key (field) and for any specific
status. It also refactors the waiters module to have one function
which is well documented and uses constants.

Change-Id: I4416addc147e496a1c0023329b606da43e7f1311