Add proxy_jump host to topology nodes discovery
When discovering nodes, this allow to use proxy jump SSH
connection info for proxy jump node if it is also a cloud node.

This is typically required when connecting to a cloud
made from vagrant instances. On such case to connect to SSH
vagrant instances (that are cloud nodes) will use SSH
vagrant configuration (IE localhost SSH port) instead of
having to SSH jump to the same host before.

Change-Id: I615d007c99be9cc157c2b7322bddbe6ad6d43014