Add healthmonitor to Octavia
Healthmonitor is Octavia's main way to check on the status of the
octavia backend members, which makes it an important part of Octavia.

After any disruptive operation, if any of the backend member servers
would get an error, then without the healthmonitor it will be harder
to spot that, as the octavia member servers will hold the NO_MONITOR
operating_status, and traffic to the LB might fail.

With the healthmonitor, the octavia member servers will enter an
ERROR operating_status, and so will the pool and the LB.
Also, we could recognize the problem we had with the members in an
easier way.

This patch adds the octavia healthmonitor to Octavia's pool stack.

Change-Id: I29c7888e04bcc1fe4cedfdfb572cbf657f5e711a