Change test_stop_dhcp_agent to test_restart_dhcp_agent
It should be enough to check if after restart of the dhcp agent
dnsmasq processes are actually restarted too.

There is already another test which checks if DHCP service works
fine for existing vms when DHCP agent is stopped.

This patch also refactored a bit {stop,start}_agent methods to avoid
code duplication. It also added restart_agent method which is pretty
similar to stop_agent and start_agent.

Change-Id: I261a14d5c41462169e8298fb564bc160ddbca636