Add timestamp to the NeutronNovaResponse object
It is object which represents log lines with notifications sent from
Neutron to Nova.
In some cases timestamp of such log may be useful if we want to get only
latest log line to check it.
This patch adds parsing log lines to find such timestamps and also adds
sort of the found and parsed log lines in the results.

Change-Id: Id9dc44b327453a155515a3c936879aab250ed855