Fix retry count=0 issue when creating new stack
When TOBIKO_PREVENT_CREATE is True, the retry count for
creating a stack were set to 0, causing the retry to
repeat forever until success.

This prevents it to happen by handling TOBIKO_PREVENT_CREATE
variable explicitly and avoid entering in the retry
creation loop in case it is True

This change also adds some additional log messages to
help investigating on stack creation problems.

Change-Id: I911e073c6a1dee926a6b16ea59930bc78c3f8396