Fix Tripleo reboot method
So far the Triple reboot_method tried to re-activate the affected
servers (which are supposed to be shutoff due to the compute reboot)
before they were in the SHUTOFF status (and hence no action would
occur).

This patch fixes the Triple reboot method by making nova to wait
for the servers to be on SHUTOFF status before re-activating them.

Change-Id: Ic8d71fa0bf5f08ef15a53b6e500e3905e5886d26