Use tobiko curl module in Octavia validators
So far, we used a raw ssh command which uses curl in order to make
sure the LB members are balanced (in the Octavia validators module).

This patch uses the Tobiko curl module instead, a change which
will make the maintenance of Octavia modules easier in the future.

To make the use of Tobiko curl possible, there has to be added a new
waiter method which waits until the members are reachable.

Change-Id: I98bd593422d7f7c8dde805fe0eb75293b5598dbe