Remove count_members variable from octavia tests
So far we used the class variable count_members in all Octavia modules.

Deciding how many members exist, saving it as a variable and testing by it
might create conflicts in the future and will be harder to maintain.
The LB should know how many members exist by using the Octavia API, not
by tests module variable.

This patch adds the list_members function and uses it in all Octavia
modules.

Change-Id: Id7c5f13db2098643ccbe4fe9a4d0ff08784d2307