Add workaround for the race between creating rooter and booting vm
In case when VM is booted before Neutron L3 agent will prepare router
namespace and metadata proxy inside that router, VM may not be able to
get metadata from Nova thus e.g. ssh-key will not be configured in such
case.
To workaround that bug, this patch adds check if guest OS finished
booting and if yes, check if SSH to the VM is possible. In case that SSH
is not possible, Tobiko will reboot vm as, if it was due to bug [1] then
during second boot metadata service should be available and ssh key
should be configured properly.

[1] https://bugs.launchpad.net/neutron/+bug/1813787

Related-Bug: #1813787
Change-Id: I1bd7b86e64ea7083f365ac84ebe1fff34cb52036