Add Active/standby amphora fault tests
So far Octavia QE had only one fault test which tested a failover.

This patch adds other tests which commit disruptive operations
and forces the amphora provider LB to run a failover process.
Finally, each test verifies the Octavia functionality.

The tests which are being added:
    * test_kill_amphora_agent
    * test_stop_keepalived
    * test_stop_haproxy

All of them will be run only if the Octavia topology is
ACTIVE_STANDBY.

Change-Id: I2ecab0973bc3467f4841991d6626166c319c96ac