Restructure Octavia stacks
So far the responsibility of creating the Octavia stack resources was
on the test cases.

That means that in case we wanted to create both loadbalancer, listener,
pool and 2 members, the test case had to create each on a separate
stack (5 stack in total), which led to more room for mistakes.

This patch merges OctaviaListenerStackFixture, OctaviaPoolStackFixture,
OctaviaMemberServerStackFixture, OctaviaOtherMemberServerStackFixture
into one stack, which will be defined as:
(lb_protocol + lb_algorithm + provider + ip_version + listener).

E.g. HttpRoundRobinAmphoraIpv4Listener.

This patch merges only the stacks which are being used by tobiko today.
Another patch which will contain stacks which are not being used (yet)
will be uploaded later.

Change-Id: I4fccbd822d449cd447f30974536a531fa3a8f48f