Skip test_server when not valid host for migration is found.
It has been seen that when compute nodes are low of resources,
migration test cases often fail because of an error reply
from the server that produce an exception of type
novaclient.exceptions.BadRequest with following text in its message:
  'No valid host was found'

This handle this kind of events by skipping server migration test
cases. It should therefore help improving stability of jobs
that are allocating many servers with only two compute nodes.

Change-Id: Ieb6e783c0331961e07db6eac79a3923a52c8cd82