Add a workaround for octavia resources cleanup
So far, in case that Octavia loadbalancers (both amphora and ovn
providers) have had children (Octavia resources which were associated
to them), they couldn't have been deleted using the current tobiko
cleanup commands.

This patch adds to the cleanup commands the loadbalancer deletion
before the stacks are being deleted.

A seperate BZ (#2060338) was opened against openstack-heat for not
adding the --cascade option when deleting LBs with children.

Workarounds BZ #2060338 until it is fixed.

Change-Id: I58bfd7d1d63c6a028470b895d588d47fd9f0fcc2