Bump paramiko version
Most tobiko tests are failing on RHEL9 environments due to ssh auth
issues. It was found that with newer paramiko version the issues
do not occur.

There is an open issue[1] in Paramiko starting from 2.9.2 (which allows
to connect to new servers like the one of RHEL 9) but in the while it
brakes the support for the old CirrOS image server (image version 5.2).

To keep support for old version we now uses disable_algorithms[1] option
when creating an SSH connection to CirrOS images. The workaround has
been documented in the Paramiko project page [2]

[1] https://github.com/paramiko/paramiko/issues/1961
[2] https://github.com/rhevm-qe-automation/python-rrmngmnt/pull/149/files#diff-7b3ed02bc73dc06b7db906cf97aa91dec2b2eb21f2d92bc5caa761df5bbc168f

Change-Id: I301c18a832a05ddfd331bddd7ad2bc839205ad2d