Fix Octavia fault tests timeouts when SSH
Octavia fault tests have failed because every failover removed the
amphora, and then the old amphora's floating ip became stale (not
attached to any resource).

Every test which used the (stale) Amphora's floating ip, received a
timeout when tried to SSH it.

This patch attaches the floating ip to the new amphora (if the Octavia
topology is ACTIVE_STANDBY, the floating ip will be attached to the
new master amphora, and if the octavia topology is STANDALONE, the
floating ip will be attached to the new only amphora).

Deleting the tobiko stacks will delete the floating ip (even if the
floating ip was set to many deleted amphoras before).

Change-Id: I76f35415fc87105cbd211b9e57ca019931a9c20d