Make undercloud keystone credentials to be working with clouds.yaml
Tobiko in generall can use clouds.yaml file and configured cloud name
to get credentials to the cloud.
But in the undercloud module, it always tried to get credentials using
environment variables and that didn't worked well with
OS_CLOUD variable.
Now when OS_CLOUD is in the env variables, it will use CloudsFile
credentials to make it working fine.

Change-Id: Ia9d976af4fd2107d22de585bf9bbdb148415c15f