Avoid setting up SSH connection when initializing it
When testing Nova server creation distruption tobiko ask
CirrOS server to close its SSH client connection. But while
asking for an CirrOS server stack registers an SSH connection
to it to ensure the connection is of CirrOS class.
This has the side-effect to make the connection manager to
estabilish the connection.

By doing so, in case the target VM has been deleted or is
shutoff just by asking for an SSH client to it will fail
while attempting to connect to it.

To avoid this and similar kind of issues the connection manager
will not setup a connection any more when one is being initialized.

Change-Id: I252cc0e6cd9eeef882e2d319147f92ca8e956fbe