Fix octavia services test on jobs with networker nodes
On OSP deployments that use networker nodes, such as composable or
upgrade, some Octavia services run on the networker nodes instead
of the controller nodes.

So far, the method _start_octavia_main_services tried to start all
octavia services for each node, because of a typo.

This patch fixes the issue by making sure that for each node, only the
stopped Octavia services will be rerun.

Change-Id: Icaa50052435628d0c977c8acf9d183ec9312479a