Refactor Ubuntu image stack
So far users could name the Ubuntu interface in any way they wanted.

The ability to name Ubuntu interface allowed tobiko to run tests
successfully for each OSP version, but it required changing it every
time a different version is being run.

This issue was fixed by setting the Ubuntu interface name in the tobiko
code.

This patch removes the unnecessary remaining parts.

Change-Id: I0fcb4b9086bf8bbae5d98983971650a2244e52a1