Configure limit of opened files for the test runner process
For some reason, when Tobiko retries many times ssh connection
and logs a lot of data in the log file, pytest may fail with
"Too many open files" exception raised.
To avoid that lets try to configure soft ulimit for opened files
to be as high as possible (same as hard limit).

Change-Id: I6fb8977d7909798a5ab6be499044cc74ac34dfa2