Remove use of Amphora FIP stacks
So far we used FIPs to access the Amphoras in Octavia's tests.

That means we had to maintain them for every failover, create, delete
and update of the amphoras.

As there are easier ways to access the Amphoras, this patch removes the
Amphoras' FIPs and uses instead the ssh-agent, which is the official way
to access them.

Change-Id: I6279468a9729b4a3da7bae3fc5b019ed1b88c18c