Remove test_ovs module from the faults tests for neutron
This module contained only one test "test_recreate_physical_bridge"
which was deleting physical bridges from the controller node, creating
those bridges again and ensuring that neutron-ovs-agent did
reconfiguration of that physical bridges properly and connectivity
through the bridges is restored.
Problem with this test is that it is "too much" intrusive. Physical
bridge needs to have some physical interface added to it to have
connectivity to the physical network. When bridge is deleted, such
port from the bridge is gone too. And Tobiko was adding bridge again,
but it wasn't adding interfaces again at all. That caused problems with
connectivity of the FIPs as well as some other stuff, like access to the
rabbitmq for services on that node when the same bridge was used for the
control plane networks.
The problem is that it would be very hard for tobiko to restore all the
configuration of the bridges as it can be some vlan interace (like
vlan50 in many Tripleo based virtual deployments used in CI), some
physical NIC or patch port to other bridge (like to br-infra in the u/s
CI configurations).

Taking all of that into account I think that it's better to simply
remove that test which in fact don't have that much value and make
Tobiko test suite more stable in testing other stuff.

Change-Id: I1f87cd74936809fe391cd8841d29ece46f9c7afe