Use unsafe option in the pre-commit's check-yaml hook
Pre-commit's check-yaml hook by default loads all yaml files to
check them. This works fine so far but as we want to have zuul
job to mirror Tobiko code to the github.com we will need to use
"secret" with custom tag "!encrypted/pkcs1-oaep". That is causing
failure of the check-yaml hook while loading yaml files.

To avoid that failure, lets use "--unsafe" option in the check-yaml
hook. With this option hook instead of loading the files,
simply parse them for syntax.

Change-Id: I9a27da6c096bae933bc9254fbad5fc5c22269560