Fix the problem with ss command line output parser
In OSP versions below 16.2 there is no 'Processes' column in the
header of `ss` command line tool output. While processes themselves
exist in the output body it still failed to parse results of the
command execution.

As a solution we may provide expected header as a parameter to the
parser as the output itself is identical for all versions

Change-Id: I2896cdbd2aa47e7bbb4d793b746a8abe0cc019ca