Add skippers to octavia-amphora ssh fault tests
Bugzilla https://bugzilla.redhat.com/show_bug.cgi?id=2126055
makes different amphoras (either Master or Backup) receive and handle
the requests whenever they reach the LB VIP, after a failover.

There might be other reasons for the tests failure, but the BZ makes it
harder to investigate, as it interupts Octavia's behavior.

This patch adds skippers for the tests which are affected from that BZ.

Change-Id: I0287e23a18f21e5b84ea4e5edc9e48bb53dceb44