Fix Octavia Active standby skipper
Octavia deploys 2 amphoras for each LB when Active standby topology is
configured.

Octavia does not provide an API to get the topology in use, and some
tests should be run only when Active standby topology is configured, so
usually in order to get the topology being used, one can get the length
of the amphorae (amphoras) per LB.

The Active standby skipper (which decides if a test should be skipped
by counting the amphoras per LB) was defined incorrectly.

This patch fixes it.

Change-Id: I6a76b67b568dfe882161049cbc6b1f89441aac4f