Shelve name of tests using a heat stack
Tobiko workers can share heat stacks. A test checks whether the stack it
requires already exists and if not, it creates it.
Sometimes, when a test fails, it deletes its stack, affecting other
tests using the same stack, running in parallel on different workers.

This patch saves persistently the name of the tests using a heat stack
using sets.
The test names are added to those sets when a new test is going to use
the stack and are removed from the sets when the test stops using the
stack (or when the test ends).
The cleanup method only deletes the stack if the numnber of tests using
it is zero.

The shelves are cleaned up everytime tobiko is executed with pytest (or
tox)

Change-Id: I92655be072efe8ecb7993c7dbf76ba930b6d9a89