Add tox max version
So far Tobiko had only a minimum tox version.

But since version 4, tox had some issues with its variables settings.
For example, in Octavia faults workflow - we set tox to run a specific
path and for some reason it is changing when the octavia_faults tox
variable is used.

The same happens with other components sanity tests (it runs the full
sanity suite) etc.

This patch adds a max version to tox, so we won't use version 4 at least
until it will be fixed.

Change-Id: I35d8e5ea2c31afac379585fb6fb9ccbbd2ae4bfc