Remove RuntimeMaxSec from iperf systemd configuration
Ubuntu images are modified to run iperf as a systemd service when a VM
starts. The parameter RuntimeMaxSec was included in that configuration
with the value of one hour.
In the unlikely event that the BW limit test was executed while the
iperf service was restarted, this could affect the test and make it
fail.
This is unlikely but it has been reproduced. During some iperf
intervals, the measured BW was 0, while 1 Mbps was expected.

Change-Id: I4622d12ede20e50344a4347da96f88e06a96f456