Improve faults tests' neutron agent health checks
One of the verifications part of the cloud health checks executed before
and after the disruptions from the faults tests was to check that all
the neutron agents are alive.
This patch adds an extra verification to this: the agents have to be
consistently alive along time.

Besides that, the number of reruns from the faults test
test_controllers_shutdown is set to 0 because it makes analysing its
failures more complicated. This test will still be flaky because we need
to determine whether there are more issues with it.

Change-Id: I354c66453493339622f99c0d18e1ff98f9f609e0