Catch some exceptions when novaclient.list_services is run
Tobiko faults tests check the list of nova services after the test
performs some disruption on the cloud. The list_services request could
fail after the disruption while the cloud is recovering and the test
should not fail due to this, but retry some times until the cloud is
healthy again.

Change-Id: I540cf2c39ef0d6cbc3cd53e917b7652ddfea5a57