Do not upgrade virtualenv to the latest version
virtualenv is a requirement for tox, so it is installed when tobiko
roles install tox.
Besides, virtualenv is also defined in the openstack requirements
upper-constraints.txt file, specifying which one is the virtualenv
valid upper version.
Due to this, virtualenv is removed from the task that upgrades other
packages as part of the tobiko-ensure-python3 role.

Change-Id: I41b76d7041face19635c9fd35e773b9144d25c30