Remove Fedora, Centos and RHEL from tobiko tests
The scenario tests were not really executed because they were marked as
flaky.
The benefit from using this images in some functional tests is irrelevant
compared with the impact in job duration and stability.
Due to these reasons, it has been decided to use only cirros and ubuntu
(for the test that required advanced images).

Change-Id: I168b1f071b3a77eb0b24586fd403d12165db1db2