Fix Octavia validators error about missing member
So far, some validations in Octavia might have run too early in the
validation flow, for example, a request is being sent to a member which
is still not ready.

This patch replaces the simple assertEqual, with an exception which will
be raised (RoundRobinException), in case that one member is not ready
yet.

The validator method caller wraps the call with a try/except clause, to
make sure Octavia has enough time to launch a ready-vm.

Change-Id: Ie3821441ff1659b7f919a34e517c32b552eaf544