Add upper-constraints file to the RTD requirements file
It's needed to cap our dependencies also on the RTD when documentation
is build to not use newer versions of packages then should be used.

For example now documentation builds on the RTD was failing because of
urllib3 2.0 which isn't compatible with Tobiko (tobiko.http._session
module in particular).
Module urllib3 is capped to 1.26.8 version in our upper-constraints.txt
file but that wasn't used in RTD builds until now.

Change-Id: I279767393716371dd16afe801c66814acd364171