Revert "Regenerate HTML report files after any test case execution"
This reverts commit 1953b9df40adf208cffdfc139f993c64a7f9261b.

Reason for revert: Due to this patch, Tobiko prints the same
Traceback several times when a test fails. More specifically, this
happens when other tests are executed after a test failure. Actually,
the Traceback is repeated as many times as tests are run after the
failure. Obviously, this makes the analysis of the test failures more
complicated.

Change-Id: I109e37dd3037d67d29cbc07b58d809f935d224d8