Improve reset_all_compute_nodes method
faults/ha tests use this method to reboot compute nodes, either
sequentially or in parallel.
In order to verify that the reboot was completed, this method executed
a 'hostname' command on the rebooted computes.
This didn't really guarantee that the reboot had been completed.
Replacing this 'hostname' command by the existing
RebootHostOperation.wait_for_operation method that obtains the output
from the 'uptime' command and compares it from the previous value: the
reboot is done when the uptime value has been updated.

Change-Id: I4f6a68eabec509b36e915fec7f4dfbd763c343d4