Re-execute podman client setup commands in case user has changed
In some cases, the user that connects to the overcloud nodes running
podman containers may change during a CI job execution (e.g. in some
upgrade jobs). When Tobiko is executed before that change, it prepares
the podman.socket service to allow connections using the old user; then,
when Tobiko is executed with the new username, the podman.socket does
not allow connections.
With this patch, Tobiko checks whether the podman.socket service is
configured with the proper username and, if not, changes its
configuration and restarts it.

Change-Id: I8466c86c886c983734133a7b4d8435073e309a59