Fix test_ipvX_subnet_nameservers when trunk ports are used
When a VM is created with trunk ports, the DNS nameservers are added
to both the parent port and the subport.
This patch checks whether the VM has a VLAN interface (i.e. it uses
trunk ports) and, in that case, tests test_ipv4_subnet_nameservers and
test_ipv6_subnet_nameservers check nameservers from both parent port
and subport subnets.

Depends-On: I6fa5b70558e886b17c63dab61fad3c7a2728d148

Change-Id: I453be8aa13a5c27d425435160756d58b672d69d6