Implement background http ping running from the guest VM
This patch adds support for running background http ping from the guest
VM. This is useful to test in background if new connections can be
established for whole time.

Background ping is implemented using very simple bash script which is
using `curl` to make requests to the destination server. The reason why
it is implemented in that way rather then using "tobiko http ping"
command already used when it is running from host or from POD (for N-S
tests) is the fact that it would require pretty many dependencies to be
installed in the guest OS to install Tobiko there. Doing it with simple
bash script is much easier and more efficient in terms of e.g. amount of
data to be downloaded during the test preparation.

Closes: #TOBIKO-132
Change-Id: I3e7062205d973318d520d98cca663d8d9fd4bf01